python/requirements.txt
asn1crypto==0.24.0
bson==0.5.8
cffi==1.12.3
cryptography==2.7
msgpack-python==0.5.6
neo4j==1.7.4
neobolt==1.7.13
neotime==1.7.4
py==1.8.0
py-bbclib==1.4.4
pycparser==2.19
pyOpenSSL==19.0.0
pytest==3.1.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
